requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.29.0
diffusers>=0.18.0
accelerate>=0.20.0
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.0
moviepy>=1.0.3
requests>=2.31.0
tqdm>=4.65.0
matplotlib>=3.7.0
pyyaml>=6.0
ffmpeg-python>=0.2.0
scikit-image>=0.20.0
einops>=0.6.0
huggingface-hub>=0.15.0
python-dotenv>=1.0.0 